Path making
Makes full paths.
Cam won't overwrite config with defaults if it somehow loads defaults

+++ camera_lib/cam_config.py
@@ -9,6 +9,7 @@ if "camera_lib" not in path:
 cfg_path = path + "/config/config.cfg"
 cfg_backup = path + "/config/config.backup"
 cfg.read(cfg_path)
+using_defaults = False
 
 # If we don't have anything, try the backup.
 if "Settings" not in cfg and os.path.isfile(cfg_backup):
@@ -17,6 +18,8 @@ if "Settings" not in cfg and os.path.isfile(cfg_backup):
 # And, if we somehow still don't have things...
 
 if "Settings" not in cfg:
+    print("WARNING: File not loaded. Using defaults")
+    using_defaults = True
     # We need our default settings.
     cfg["Settings"] = {}
     cfg["Settings"]["JPEG"] = "True"
@@ -33,8 +36,12 @@ if "Info" not in cfg:
 
 def save_config():
     # Deletes the old backup, saves our current as the backup, then saves the new version.
-    if os.path.isfile(cfg_backup):
-        os.remove(cfg_backup)
-    os.rename(cfg_path, cfg_backup)
-    with open(cfg_path, 'w') as cfgfile:
-        cfg.write(cfgfile)
+    if using_defaults:
+        # Don't overwrite the files, just in case.
+        pass
+    else:
+        if os.path.isfile(cfg_backup):
+            os.remove(cfg_backup)
+        os.rename(cfg_path, cfg_backup)
+        with open(cfg_path, 'w') as cfgfile:
+            cfg.write(cfgfile)
